+++ NEWS.md
@@ -5,6 +5,9 @@
 - `txt_style()` now accepts a `class` argument specifying a vector of classes to
   be applied to the `<span>` of the decorated text (@gadenbuie, #18).
 
+- `chunk_addin()` RStudio addin have been added. add `decorate()` chunk after
+selected chunk when used.
+
 ## Bugs and fixes
 
 * NULL document types are now treated as default html